remove gen-docs as a command and make it a normal script
