+++ package.json
@@ -10,7 +10,7 @@
     "postpublish": "pinst --enable",
     "build": "tsc --project tsconfig.json && ./shebangify.js ./dist/index.js && chmod +x ./dist/index.js",
     "clean": "rm -rf dist/",
-    "gen-docs": "yarn build && ./dist/index.js gen-docs && git add COMMANDS.md",
+    "gen-docs": "ts-node ./gen-docs.ts && git add COMMANDS.md",
     "prettier": "pretty-quick --staged"
   },
   "repository": {
@@ -49,6 +49,7 @@
     "pinst": "^2.1.6",
     "prettier": "^2.3.1",
     "pretty-quick": "^3.1.0",
+    "ts-node": "^10.1.0",
     "typescript": "^4.3.3"
   },
   "dependencies": {
